+++ src/Exporter.php
@@ -6,11 +6,18 @@ namespace Bakame\Stackwatch;
 
 interface Exporter
 {
-    public function exportMetrics(Metrics $metrics): void;
-    public function exportReport(Report $report): void;
     public function exportEnvironment(Environment $environment): void;
-    public function exportStatistics(Statistics $statistics): void;
     public function exportSnapshot(Snapshot $snapshot): void;
+    public function exportMetrics(Metrics $metrics): void;
+    public function exportSpan(Span $span): void;
+    public function exportStatistics(Statistics $statistics): void;
+    public function exportReport(Report $report): void;
+    /**
+     * @param (callable(Span): bool)|string|null $label
+     */
     public function exportSpanAggregator(SpanAggregator $spanAggregator, callable|string|null $label = null): void;
-    public function exportTimeline(Timeline $timeline): void;
+    /**
+     * @param ?callable(Snapshot): bool $filter
+     */
+    public function exportTimeline(Timeline $timeline, ?callable $filter = null): void;
 }