+++ src/Ide.php
@@ -30,14 +30,24 @@ enum Ide: string
     public function uri(CallLocation $location, ?string $project = null): string
     {
         return match ($this) {
-            self::JetBrains => 'jetbrains://php-storm/navigate/reference?'. http_build_query(['project' => $project, 'path' => $location->path, 'line'=> $location->line]),
+            self::JetBrains => 'jetbrains://php-storm/navigate/reference?'.http_build_query(['project' => $project, 'path' => $location->path, 'line' => $location->line]),
             self::VSCode => 'vscode://file/'.ltrim((string) $location->path, '/').':'.$location->line,
-            default => 'phpstorm://open?'.http_build_query(['file' => $location->path, 'line'=> $location->line]),
+            default => 'phpstorm://open?'.http_build_query(['file' => $location->path, 'line' => $location->line]),
         };
     }
 
     public function path(CallLocation $location): string
     {
         return $location->path.':'.$location->line;
+    }
+
+    public function link(CallLocation $location, ?string $project = null, AnsiStyle ...$styles): string
+    {
+        $path = $this->path($location);
+        if (Environment::current()->isCli()) {
+            return $path;
+        }
+
+        return '<a class="'.AnsiStyle::inlineClasses(...$styles).'" href="'.$this->uri($location, $project).'">'.$path.'</a>';
     }
 }